momentum/requirements.txt
streamlit
pandas
mariadb
toml
yfinance
mplfinance
python-dateutil